SoC: Remove ftdi_device_desc in openocd cfg
For RVSTARv2 and HummingBird Debugger v2, the USB
JTAG description is changed from "Dual RS232-HS"
to "USB <-> JTAG-DEBUGGER", to be compatiable with
previous version, we decided to remove this ftdi_device_desc
in openocd cfg, so different version of JTAG could be
recognized.

If you need to specify a JTAG device, you can add this ftdi_device_desc
by yourself.

Signed-off-by: Huaqi Fang <578567190@qq.com>