OS: bugfix on stack sp alignment for rtos porting
see https://github.com/riscv-non-isa/riscv-elf-psabi-doc/blob/301374e92976e298e676e7129a6212926b2299ce/riscv-cc.adoc?plain=1#L245-L259

The stack grows downwards (towards lower addresses) and the stack pointer shall be aligned to a 128-bit boundary upon procedure entry.

The ILP32E calling convention is designed to be usable with the RV32E ISA. This calling convention is the same as the integer calling convention, except for the following differences. The stack pointer need only be aligned to a 32-bit boundary.

This bug may appear when you declare a struct variable in a function with alignment requirement, when access to this variable,
such as store value to it, it may corrupt the saved registers on stack such as ra, then you may return to a wrong pc

eg.

struct urequest
{
    uint8_t  request_type;
    uint8_t  bRequest;
    uint16_t wValue;
    uint16_t wIndex;
    uint16_t wLength;
};

void do_something(void)
{
    volatile __ALIGNED(32) struct urequest setup;

    /* make use of this setup variable */
}

Signed-off-by: Huaqi Fang <578567190@qq.com>